Move components to specific folder
