+++ client/src/components/columnPicker/index.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { shallow } from "enzyme";
-import ColumnPicker from "./columnPicker";
+import ColumnPicker from "./index";
 
 describe("ColumnPicker component", () => {
   it("Renders without crashing", () => {
