+++ client/src/components/columnPicker/index.tsx
@@ -1,8 +1,8 @@
 import { ControlGroup, FormGroup, InputGroup, Switch } from "@blueprintjs/core";
 import * as React from "react";
 
-import StringSelect from "./selects/stringSelect";
-import { ISourceColumn, ISourceSchema } from "../types";
+import StringSelect from "../selects/stringSelect";
+import { ISourceColumn, ISourceSchema } from "../../types";
 
 export interface IProps {
   ownerChangeCallback?: any;
